requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
certifi==2025.11.12
gunicorn==21.2.0
stripe==7.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Pillow==10.1.0
paynow==1.0.5
africastalking==1.2.8
